allow local administrator extensions to the back-end config
Currently the content of the /etc/crypto-policies/back-ends
files are entirely controlled by the update-crypto-policies
script. This is somewhat restrictive, for example, with
gnutls the script only knows about setting the "SYSTEM"
policy and is unable to set any other application specific
policies.

This patch introduces support for local administrator defined
extensions, to augment the content written to the back-end
files. These are kept in the local.d directory in the back-end
specific config format.

For example, with gnutls the administrator can now do

 echo "LIBVIRT=NONE:+VERS-TLS-ALL" >>
   /etc/crypto-policies/local.d/gnutls.config

And when running update-crypto-policies, this will be appended
to the /etc/crypto-policies/back-ends/gnutls.config file.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>